Fix x64 clobber offsets with unwind info
The stack offsets provided to UnwindInst::SaveReg instructions are
relative to the beginning of the clobber area of the stack, and on main
the x64 backend computes those offsets by assuming a SP-relative offset
and then adjusting it. This is problematic, as any changes to the frame
layout might require updates in two locations: where we compute the
initial SP offset, and where we compute the adjustment to make the
offset relative to the clobber area again.

This PR fixes this issue by making the offset we track in the loop
always relative to the beginning of the clobber area, and adding the
offset from SP only when emitting the store. This relies on the
assumption that the clobber area is sized according to the constraints
of the registers stored within it (16-byte alignment is required for
floating point registers) but this assumption is enforced by existing
behavior in compute_clobber_size.

This also fixes a long-standing bug with floating point register spill
offsets, as the offset was saved before the alignment to a 16-byte
offset was made, leaving the resulting unwind info potentially pointing
at the wrong offset.