Also modify gen_clobber_restore to load positive offsets from SP
This fixes a bug: we're not supposed to write below SP, as a signal
handler is allowed to clobber anything below it. Writing our clobbers to
negative offsets of SP leaves a window open where they might be
clobbered if we receive a signal while executing the prologue or
epilogue.