x64: Remove unnecessary `put_in_xmm` in `x64_ucomis`
This commit is a result of [discussion on Zulip][Zulip] which concluded
that there's no longer any need for these `put_in_xmm` methods. The
condition these were previously protecting against is now upheld by
support added in #4061

I've additionally added a number of new tests here to assert that load
sinking isn't happening where it's not supposed to, and for this commit
I've manually verified it's all as expected.

[Zulip]: https://bytecodealliance.zulipchat.com/#narrow/stream/217117-cranelift/topic/x64.2C.20fcmp.2C.20and.20load-op.20merging/near/433319056