x64: Add AVX-encoded versions of `ucomis{s,d}` and `ptest` (#8409)
* x64: Remove unnecessary `put_in_xmm` in `x64_ucomis`

This commit is a result of [discussion on Zulip][Zulip] which concluded
that there's no longer any need for these `put_in_xmm` methods. The
condition these were previously protecting against is now upheld by
support added in #4061

I've additionally added a number of new tests here to assert that load
sinking isn't happening where it's not supposed to, and for this commit
I've manually verified it's all as expected.

[Zulip]: https://bytecodealliance.zulipchat.com/#narrow/stream/217117-cranelift/topic/x64.2C.20fcmp.2C.20and.20load-op.20merging/near/433319056

* x64: Change signature of `x64_ucomis` to take registers

Previously this was the only instruction helper which took a `Value` as
an argument. Now it works like all other instruction helpers by taking a
typed register as an argument, namely `Xmm` and `XmmMem`.

* x64: Add AVX encodings of float-compare instructions

This fills out the final remaining AVX-related instructions that
Cranelift can emit, so now Cranelift should use AVX instructions for all
float/simd operations when AVX is enabled.

* Fix test expectation