x64: Use r10 for the indirect tail call destination (#8407)
* Always use r10 for indirect return calls, and indirect winch calls

Use `r10` for the destination of indirect return calls, and indirect
calls using the winch calling convention, as it is a caller-saved
register.

For tail calls, this ensures that we won't accidentally pick a
callee-saved register for the destination, clobbering it when we restore
callee-saves in the call to `emit_return_call_common_sequence`.

For winch calls, using `r10` instead of `r15` means that it's still
possible to use the pinned register in combination with winch.

* Always use `r11` for the temp in return_call

* Switch ReturnCallUnknown to taking the callee as a Reg