cranelift/x64: Narrow `test` immediate operands
The x86 `test` instruction does a bitwise-and operation, setting flags,
but otherwise discarding the result. And when one operand is an
immediate constant, any zero bits in the constant can't have an effect
on the result. So we can emit shorter versions of this instruction by
truncating the most significant zero bits and using the narrowest
possible immediate form for the provided constant.