x64: Swap operand order of `cmp` in internal representation (#8408)
* x64: Rename fields of `CmpRmiR` and swap them

This commit is the start of work that is the equivalent of #8362 but for
comparisons of general purpose instructions. Here only the instruction
arguments are swapped and renamed. All callers preserve the
right-to-left ordering and a subsequent commit will swap them to
left-to-right.

* x64: Swap `test_rr` operands in Winch

* x64: Swap operands of `cmp_rr` in Winch

* x64: Swap operands of `cmp_ir` in Winch

* x64: Swap operands for the `cmp` masm method in Winch

This additionally needed to affect the `branch` method. Methods now
additionally document what's the left-hand-side and what's the
right-hand-side.

Of note here was that the `branch` instruction actually used the terms
"lhs" and "rhs" incorrectly where the left/right-ness was actually
swapped in the actual comparison. This causes no issues, however, since
`branch` was only ever used when both operands were the same or a
reflexive condition was used.

* x64: Swap operands for `Inst::cmp_rmi_r`

* x64: Swap operand order of `cmp_rmi_r` ISLE helper

Also update all callers to swap orders as well.

* x64: Swap operand order of `x64_test` helper

* Swap operand order of `x64_cmp_imm` helper

* x64: Swap operand order of `cmp`

* x64: Define `x64_cmp_imm` with `x64_cmp`

Minor refactoring which clarifies that this is just a normal `cmp`
except with a different signature.

* x64: Use `x64_cmp_imm` in a few more locations

A bit easier on the eyes to read.