cranelift: Don't recompute allocatable register set
The implementation of `PRegSet::from(self.vcode.machine_env())` does a
surprising amount of work: it lazily initializes a `OnceLock` in the
backend, then loops over six vectors of registers and adds them all to
the `PRegSet`.

We could likely implement that better, but in the meantime at least we
can avoid repeating that work for every single machine instruction. The
`PRegSet` itself only takes a few words to store so it's cheap to just
keep it around.

I discovered this because when the call to `self.vcode.machine_env()` is
in the middle of the loop, that prevents holding a mutable borrow on
parts of `self.vcode`, which I want to be able to do in another PR.