cranelift: Simplify computing the clobber set (#8469)
We have a perfectly good data structure for holding sets of physical
registers. We don't need a hash-set to determine whether we already
added a register to an array when we can just add it to a PRegSet and
then extract a suitable array out of that.

This change also makes the computed array sorted, which isn't
particularly important since the backends all modify it and then sort it
again, but is generally a nice property to have.