Gate type-builder types from `wasmtime-environ` on `compile` (#8484)
* Gate type-builder types from `wasmtime-environ` on `compile`

This commit gates the `*Builder` types related to building sets of types
in the `wasmtime-environ` crate on the `compile` feature. This helps
bring in less code when the feature is disabled and helps exclude some
dependencies for the upcoming `no_std` migration as well.

This commit doesn't change anything, it's just moving code around.

* Remove no-longer-needed import

prtest:full