cranelift: Split predecessor/successor lists
This establishes the property that the VCode's various lists of ranges
each fully cover the index range of another list. Previously, the
block_succ_range list covered the first half of block_succs_preds, and
the block_pred_range list covered the second half.

While I was in the area, I replaced the O(n log n) sort in
compute_preds_from_succs with a linear-time counting sort, which uses
less temporary storage and directly computes the ranges we want as a
byproduct.