cranelift: Rewrite operands with allocations in-place
This makes all other code that uses AllocationConsumer into no-ops
because the list of allocations in all other places are now always
empty.

There is a bunch of dead code still left over after this. This part is
just the stuff I can delete without changing function signatures
everywhere. The various functions which are now no-ops are called in a
lot of places and cleaning those up will take more careful review.