wasi-nn: factor out common test-program code
`wasi-nn`'s test program suite is light at the moment but, in order to
expand it, this change factors out some of the common bits that are
being used in the `test-programs` crate. Since all of the tests perform
some kind of image classification, the new `nn` module gains `classify`
and `sort_results` functions to help with this exact case.

prtest:full