wasmtime: Check stack limits only on exit from wasm
Currently, we do an explicit check for stack overflow on entry to every
WebAssembly function. But that costs some time, and is a significant
performance hit for very short functions.

This commit instead switches Wasmtime to relying on guard pages at the
end of the stack to catch stack overflow, so the MMU does this check for
"free". This means we may allow deeper recursion in guest code than we
did before.

To make this work, we need Wasmtime's signal handlers to recognize when
a guest memory fault is in a stack guard page and report the appropriate
stack-overflow trap code.

Note that we can't turn host-code signals into guest traps, so the
signal handlers have to verify that the signal occurred in guest code.

When the guest calls host code (explicitly due to calling an imported
host function, or implicitly due to a libcall inserted by Wasmtime or
Cranelift), we also need to ensure that there is enough stack space
available for the host code to not hit the guard pages. We do that by
checking the stack limit that the embedder provided in the trampolines
where we exit wasm.