cranelift: Rewrite operands with allocations in-place (#8524)
* cranelift: Rewrite operands with allocations in-place

This makes all other code that uses AllocationConsumer into no-ops
because the list of allocations in all other places are now always
empty.

There is a bunch of dead code still left over after this. This part is
just the stuff I can delete without changing function signatures
everywhere. The various functions which are now no-ops are called in a
lot of places and cleaning those up will take more careful review.

* Make OperandVisitor implementable with a closure

Instead of extending AllocationConsumer to update operands in place,
introduce a new type and delete the rest of the internals of
AllocationConsumer.

This involves deleting the lifetime annotations from all uses of
AllocationConsumer. I could have kept a `PhantomData` or something
around but didn't feel like bothering.

* Debug-assert that AllocationConsumer has no allocations to consume

This is what justifies deleting all this code, so let's go ahead and
assert it until the code is actually gone.