V8 differential fuzzing: add new table out-of-bounds error message to expectations list. (#8564)
Fixes https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=68735.

That fuzzbug bisected to the call-indirect caching changes, but this
turned out to be a red herring (the options added in that PR mean that
the fuzzbug config deserializes differently prior to the commit). In
any case, it's an easy fix -- it appears that V8 added a new error
message, so we need to add it to the allowlist of messages that we
expect for a table out-of-bounds condition.