wasmtime: Remove ALL constant phis
When we're trying to delete block-params that can be replaced by a
single dominating value, we weren't handling some cases.

In particular, if we concluded that a block formal parameter (say, `v3`)
had more than one value, then we believed that any uses of that
parameter (say, defining another formal parameter `v4`) also had more
than one value, and therefore could not be deleted.

However, in such cases we can try using `v3` itself as the definition of
`v4`. If there are no other definitions of `v4`, then it can be deleted.

With this change, if a block has only one predecessor, it is now true
that this pass will delete all of its block params. We couldn't rely on
that property before.